Add 13.2.3

+++ Frameworks/AssetsLibrary.framework/ALAssetRepresentation.h
@@ -13,11 +13,11 @@
 + (void)_enableImageDataUsesMap;
 
 - (struct CGImage { }*)CGImageWithOptions:(id)arg1;
-- (struct CGImage { }*)CGImageWithOptions:(id)arg1 format:(int)arg2;
+- (struct CGImage { }*)CGImageWithOptions:(id)arg1 format:(unsigned short)arg2;
 - (id)UTI;
 - (int)_fileDescriptor;
 - (id)_imageData;
-- (struct CGImage { }*)_largeDisplayableImageForFormatID:(int)arg1;
+- (struct CGImage { }*)_largeDisplayableImageForFormatID:(unsigned short)arg1;
 - (void)dealloc;
 - (id)description;
 - (struct CGSize { double x1; double x2; })dimensions;
