+++ Frameworks/MediaPlayer.framework/_MPMediaLibraryEntityChange.h
@@ -4,18 +4,18 @@
 
 @interface _MPMediaLibraryEntityChange : NSObject {
     NSString * _anchor;
-    int  _deletionType;
+    long long  _deletionType;
     MPMediaEntity * _entity;
 }
 
 @property (nonatomic, readonly, copy) NSString *anchor;
-@property (nonatomic, readonly) int deletionType;
+@property (nonatomic, readonly) long long deletionType;
 @property (nonatomic, readonly) MPMediaEntity *entity;
 
 - (void).cxx_destruct;
 - (id)anchor;
-- (int)deletionType;
+- (long long)deletionType;
 - (id)entity;
-- (id)initWithEntity:(id)arg1 anchor:(id)arg2 deletionType:(int)arg3;
+- (id)initWithEntity:(id)arg1 anchor:(id)arg2 deletionType:(long long)arg3;
 
 @end
