+++ PrivateFrameworks/DCIMServices.framework/DCIMImageWriter.h
@@ -17,7 +17,7 @@
 + (void)writeableDataForImageData:(id)arg1 imageUTIType:(struct __CFString { }*)arg2 thumbnailDataOut:(id*)arg3 imageUTITypeOut:(id*)arg4 exifPropertiesOut:(id*)arg5 isJPEGOut:(bool*)arg6 imageDataOut:(id*)arg7;
 
 - (id)_cameraAssetExtensionForType:(short)arg1;
-- (void)_writeJPEGFromIOSurface:(struct __IOSurface { }*)arg1 toPath:(id)arg2 orientation:(int)arg3;
+- (void)_writeJPEGFromIOSurface:(struct __IOSurface { }*)arg1 toPath:(id)arg2 orientation:(long long)arg3;
 - (bool)_writeJPEGToURL:(id)arg1 withData:(id)arg2 thumbnail:(id)arg3 properties:(id)arg4 duringBurst:(bool)arg5;
 - (int)copyPrimaryFormatImageData:(id)arg1 toURL:(id)arg2 properties:(id)arg3;
 - (bool)saveImageJobToDisk:(id)arg1;
