+++ PrivateFrameworks/PersonalizationPortraitInternals.framework/PPContactDiskCache.h
@@ -3,7 +3,7 @@
  */
 
 @interface PPContactDiskCache : NSObject {
-    double  _lastCreatedAt;
+    long long  _lastCreatedAt;
 }
 
 + (id)sharedInstance;
