+++ PrivateFrameworks/ProactiveML.framework/PMLTrainingDatabaseMigrations.h
@@ -6,16 +6,16 @@
     NSDictionary * _migrations;
 }
 
-@property (nonatomic, readonly) unsigned long long maxVersion;
+@property (nonatomic, readonly) unsigned int maxVersion;
 @property (nonatomic, readonly) NSDictionary *migrations;
 
 + (id)mockMigrationsByAddingQueries:(id)arg1;
-+ (id)skipFromZeroSchema:(unsigned long long*)arg1;
++ (id)skipFromZeroSchema:(unsigned int*)arg1;
 
 - (void).cxx_destruct;
 - (id)init;
 - (id)initWithMigrations:(id)arg1;
-- (unsigned long long)maxVersion;
+- (unsigned int)maxVersion;
 - (id)migrations;
 
 @end
