+++ PrivateFrameworks/ScreenReaderCore.framework/SCRCMathFencedExpression.h
@@ -25,7 +25,7 @@
 - (id)openString;
 - (void)setCloseString:(id)arg1;
 - (void)setOpenString:(id)arg1;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(bool)arg2;
-- (id)speakableSegmentsWithSpeakingStyle:(int)arg1 upToDepth:(unsigned long long)arg2 treePosition:(id)arg3;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(bool)arg2;
+- (id)speakableSegmentsWithSpeakingStyle:(long long)arg1 upToDepth:(unsigned long long)arg2 treePosition:(id)arg3;
 
 @end
