+++ PrivateFrameworks/ScreenReaderCore.framework/SCRCMathNumberExpression.h
@@ -6,6 +6,6 @@
 
 - (bool)isNumber;
 - (id)mathMLString;
-- (id)speakableDescriptionWithSpeakingStyle:(int)arg1 arePausesAllowed:(bool)arg2;
+- (id)speakableDescriptionWithSpeakingStyle:(long long)arg1 arePausesAllowed:(bool)arg2;
 
 @end
